+++ pcs/lib/cib/resource/bundle.py
@@ -11,13 +11,11 @@ from pcs.lib import reports, validate
 from pcs.lib.cib.nvpair import (
     append_new_meta_attributes,
     arrange_first_meta_attributes,
+    META_ATTRIBUTES_TAG,
 )
 from pcs.lib.cib.resource.primitive import TAG as TAG_PRIMITIVE
 from pcs.lib.cib.tools import find_element_by_tag_and_id
-from pcs.lib.errors import (
-    LibraryError,
-    ReportListAnalyzer,
-)
+from pcs.lib.errors import LibraryError
 from pcs.lib.pacemaker.values import sanitize_id
 from pcs.lib.xml_tools import (
     get_sub_element,
@@ -26,7 +24,7 @@ from pcs.lib.xml_tools import (
 
 TAG = "bundle"
 
-_docker_options = set((
+GENERIC_CONTAINER_OPTIONS = frozenset((
     "image",
     "masters",
     "network",
@@ -36,11 +34,26 @@ _docker_options = set((
     "replicas-per-host",
 ))
 
-_network_options = set((
+NETWORK_OPTIONS = frozenset((
     "control-port",
     "host-interface",
     "host-netmask",
     "ip-range-start",
+))
+
+PORT_MAP_OPTIONS = frozenset((
+    "id",
+    "port",
+    "internal-port",
+    "range",
+))
+
+STORAGE_MAP_OPTIONS = frozenset((
+    "id",
+    "options",
+    "source-dir",
+    "source-dir-root",
+    "target-dir",
 ))
 
 def is_bundle(resource_el):
@@ -62,9 +75,7 @@ def validate_new(
     list of dict storage_map -- list of storage mapping options
     bool force_options -- return warnings instead of forceable errors
     """
-    report_list = []
-
-    report_list.extend(
+    return (
         validate.run_collection_of_option_validators(
             {"id": bundle_id},
             [
@@ -72,30 +83,17 @@ def validate_new(
                 validate.value_id("id", "bundle name", id_provider),
             ]
         )
-    )
-
-    aux_reports = _validate_container_type(container_type)
-    report_list.extend(aux_reports)
-    if not ReportListAnalyzer(aux_reports).error_list:
-        report_list.extend(
-            # TODO call the proper function once more container_types are
-            # supported by pacemaker
-            _validate_container_docker_options_new(
-                container_options,
-                force_options
-            )
+        +
+        validate_reset(
+            id_provider,
+            container_type,
+            container_options,
+            network_options,
+            port_map,
+            storage_map,
+            force_options
         )
-    report_list.extend(
-        _validate_network_options_new(network_options, force_options)
     )
-    report_list.extend(
-        _validate_port_map_list(port_map, id_provider, force_options)
-    )
-    report_list.extend(
-        _validate_storage_map_list(storage_map, id_provider, force_options)
-    )
-
-    return report_list
 
 def append_new(
     parent_element, id_provider, bundle_id, container_type, container_options,
@@ -115,30 +113,111 @@ def append_new(
     dict meta_attributes -- meta attributes
     """
     bundle_element = etree.SubElement(parent_element, TAG, {"id": bundle_id})
-    # TODO create the proper element once more container_types are supported
-    # by pacemaker
-    docker_element = etree.SubElement(bundle_element, "docker")
-    # Do not add options with empty values. When updating, an empty value means
-    # remove the option.
-    update_attributes_remove_empty(docker_element, container_options)
+    _append_container(bundle_element, container_type, container_options)
     if network_options or port_map:
-        network_element = etree.SubElement(bundle_element, "network")
-        # Do not add options with empty values. When updating, an empty value
-        # means remove the option.
-        update_attributes_remove_empty(network_element, network_options)
-    for port_map_options in port_map:
-        _append_port_map(
-            network_element, id_provider, bundle_id, port_map_options
+        _append_network(
+            bundle_element,
+            id_provider,
+            bundle_id,
+            network_options,
+            port_map,
         )
     if storage_map:
-        storage_element = etree.SubElement(bundle_element, "storage")
-    for storage_map_options in storage_map:
-        _append_storage_map(
-            storage_element, id_provider, bundle_id, storage_map_options
-        )
+        _append_storage(bundle_element, id_provider, bundle_id, storage_map)
     if meta_attributes:
         append_new_meta_attributes(bundle_element, meta_attributes, id_provider)
     return bundle_element
+
+def validate_reset(
+    id_provider, container_type, container_options, network_options,
+    port_map, storage_map, force_options=False
+):
+    """
+    Validate bundle parameters, return list of report items
+
+    IdProvider id_provider -- elements' ids generator and uniqueness checker
+    string container_type -- bundle container type
+    dict container_options -- container options
+    dict network_options -- network options
+    list of dict port_map -- list of port mapping options
+    list of dict storage_map -- list of storage mapping options
+    bool force_options -- return warnings instead of forceable errors
+    """
+    return (
+        _validate_container(container_type, container_options, force_options)
+        +
+        _validate_network_options_new(network_options, force_options)
+        +
+        _validate_port_map_list(port_map, id_provider, force_options)
+        +
+        _validate_storage_map_list(storage_map, id_provider, force_options)
+    )
+
+def reset(
+    bundle_element, id_provider, bundle_id, container_type, container_options,
+    network_options, port_map, storage_map, meta_attributes
+):
+    """
+    Remove configuration of bundle_element and create new one.
+
+    etree bundle_element -- the bundle element that will be reset
+    IdProvider id_provider -- elements' ids generator
+    string bundle_id -- id of the bundle
+    string container_type -- bundle container type
+    dict container_options -- container options
+    dict network_options -- network options
+    list of dict port_map -- list of port mapping options
+    list of dict storage_map -- list of storage mapping options
+    dict meta_attributes -- meta attributes
+    """
+    # pylint: disable=too-many-arguments
+
+    # Old bundle configuration is removed and re-created. We aren't trying
+    # to keep ids:
+    # * It doesn't make sense to reference these ids.
+    # * Newly created ids are based on (are prefixed by) the bundle element id,
+    #   which does not change. Therefore, it is VERY HIGHLY probable the newly
+    #   created ids will be the same as the original ones.
+    elements_without_reset_impact = []
+
+    # Elements network, storage and meta_attributes must be kept even if they
+    # are without children.
+    # See https://bugzilla.redhat.com/show_bug.cgi?id=1642514
+    #
+    # The only scenario that makes sense is that these elements are empty
+    # and no attributes or children are requested for them. So we collect only
+    # deleted tags and we will ensure creation minimal relevant elements at
+    # least.
+    indelible_tags = []
+    for child in list(bundle_element):
+        if child.tag in ["network", "storage", META_ATTRIBUTES_TAG]:
+            indelible_tags.append(child.tag)
+        elif child.tag != "docker":
+            # Only primitive should be found here, currently.
+            # The order of various element tags has no practical impact so we
+            # don't care about it here.
+            elements_without_reset_impact.append(child)
+        bundle_element.remove(child)
+
+    _append_container(bundle_element, container_type, container_options)
+    if network_options or port_map or "network" in indelible_tags:
+        _append_network(
+            bundle_element,
+            id_provider,
+            bundle_id,
+            network_options,
+            port_map,
+        )
+    if storage_map or "storage" in indelible_tags:
+        _append_storage(bundle_element, id_provider, bundle_id, storage_map)
+    if meta_attributes or META_ATTRIBUTES_TAG in indelible_tags:
+        append_new_meta_attributes(
+            bundle_element,
+            meta_attributes,
+            id_provider,
+        )
+    for element in elements_without_reset_impact:
+        bundle_element.append(element)
 
 def validate_update(
     id_provider, bundle_el, container_options, network_options,
@@ -161,7 +240,7 @@ def validate_update(
     report_list = []
 
     container_el = _get_container_element(bundle_el)
-    if container_el.tag == "docker":
+    if container_el is not None and container_el.tag == "docker":
         # TODO call the proper function once more container types are
         # supported by pacemaker
         report_list.extend(
@@ -323,12 +402,16 @@ def get_inner_resource(bundle_el):
         return resources[0]
     return None
 
-def _validate_container_type(container_type):
-    return validate.value_in("type", ("docker", ), "container type")({
-        "type": container_type,
-    })
+def _validate_container(container_type, container_options, force_options=False):
+    if container_type != "docker":
+        return [
+            reports.invalid_option_value(
+                "container type",
+                container_type,
+                ["docker"],
+            )
+        ]
 
-def _validate_container_docker_options_new(options, force_options):
     validators = [
         validate.is_required("image", "container"),
         validate.value_not_empty("image", "image name"),
@@ -337,11 +420,14 @@ def _validate_container_docker_options_new(options, force_options):
         validate.value_positive_integer("replicas-per-host"),
     ]
     return (
-        validate.run_collection_of_option_validators(options, validators)
+        validate.run_collection_of_option_validators(
+            container_options,
+            validators
+        )
         +
         validate.names_in(
-            _docker_options,
-            options.keys(),
+            GENERIC_CONTAINER_OPTIONS,
+            container_options.keys(),
             "container",
             report_codes.FORCE_OPTIONS,
             force_options
@@ -372,7 +458,7 @@ def _validate_container_docker_options_update(
         +
         validate.names_in(
             # allow to remove options even if they are not allowed
-            _docker_options | _options_to_remove(options),
+            GENERIC_CONTAINER_OPTIONS | _options_to_remove(options),
             options.keys(),
             "container",
             report_codes.FORCE_OPTIONS,
@@ -390,7 +476,7 @@ def _validate_network_options_new(options, force_options):
         validate.run_collection_of_option_validators(options, validators)
         +
         validate.names_in(
-            _network_options,
+            NETWORK_OPTIONS,
             options.keys(),
             "network",
             report_codes.FORCE_OPTIONS,
@@ -453,7 +539,7 @@ def _validate_network_options_update(
         +
         validate.names_in(
             # allow to remove options even if they are not allowed
-            _network_options | _options_to_remove(options),
+            NETWORK_OPTIONS | _options_to_remove(options),
             options.keys(),
             "network",
             report_codes.FORCE_OPTIONS,
@@ -462,12 +548,6 @@ def _validate_network_options_update(
     )
 
 def _validate_port_map_list(options_list, id_provider, force_options):
-    allowed_options = [
-        "id",
-        "port",
-        "internal-port",
-        "range",
-    ]
     validators = [
         validate.value_id("id", "port-map id", id_provider),
         validate.depends_on_option(
@@ -489,7 +569,7 @@ def _validate_port_map_list(options_list, id_provider, force_options):
             validate.run_collection_of_option_validators(options, validators)
             +
             validate.names_in(
-                allowed_options,
+                PORT_MAP_OPTIONS,
                 options.keys(),
                 "port-map",
                 report_codes.FORCE_OPTIONS,
@@ -499,13 +579,6 @@ def _validate_port_map_list(options_list, id_provider, force_options):
     return report_list
 
 def _validate_storage_map_list(options_list, id_provider, force_options):
-    allowed_options = [
-        "id",
-        "options",
-        "source-dir",
-        "source-dir-root",
-        "target-dir",
-    ]
     source_dir_options = ["source-dir", "source-dir-root"]
     validators = [
         validate.value_id("id", "storage-map id", id_provider),
@@ -519,7 +592,7 @@ def _validate_storage_map_list(options_list, id_provider, force_options):
             validate.run_collection_of_option_validators(options, validators)
             +
             validate.names_in(
-                allowed_options,
+                STORAGE_MAP_OPTIONS,
                 options.keys(),
                 "storage-map",
                 report_codes.FORCE_OPTIONS,
@@ -550,6 +623,26 @@ def _value_host_netmask(option_name, force_options):
         allow_extra_values=force_options
     )
 
+def _append_container(bundle_element, container_type, container_options):
+    # Do not add options with empty values. When updating, an empty value means
+    # remove the option.
+    update_attributes_remove_empty(
+        etree.SubElement(bundle_element, container_type),
+        container_options,
+    )
+
+def _append_network(
+    bundle_element, id_provider, id_base, network_options, port_map
+):
+    network_element = etree.SubElement(bundle_element, "network")
+    # Do not add options with empty values. When updating, an empty value means
+    # remove the option.
+    update_attributes_remove_empty(network_element, network_options)
+    for port_map_options in port_map:
+        _append_port_map(
+            network_element, id_provider, id_base, port_map_options
+        )
+
 def _append_port_map(parent_element, id_provider, id_base, port_map_options):
     if "id" not in port_map_options:
         id_suffix = None
@@ -566,6 +659,16 @@ def _append_port_map(parent_element, id_provider, id_base, port_map_options):
     # remove the option.
     update_attributes_remove_empty(port_map_element, port_map_options)
     return port_map_element
+
+def _append_storage(bundle_element, id_provider, id_base, storage_map):
+    storage_element = etree.SubElement(bundle_element, "storage")
+    for storage_map_options in storage_map:
+        _append_storage_map(
+            storage_element,
+            id_provider,
+            id_base,
+            storage_map_options,
+        )
 
 def _append_storage_map(
     parent_element, id_provider, id_base, storage_map_options